requirements.txt
fastapi >= 0.112.1
openai >= 1.41.1
pydantic==2.8.2
uvicorn[standard] >= 0.30.6
python-dotenv >= 1.0.1
langchain >= 0.2.14
langchain-community >= 0.2.12
starlette~=0.38.2
langchain-openai >= 0.1.22
pyjwt[crypto] >= 2.9.0
langchain-core >= 0.2.33
langgraph >= 0.2.41